Move EventCounter into finesse.gui module
It is only used in finesse.gui so that's probably where it should live.
