+++ finesse/gui/event_counter.py
